+++ app/src/main/java/br/com/lucasisrael/jetposemovies/moviesgenre/data/datasource/MoviesFromGenreDataSource.kt
@@ -1,6 +1,6 @@
 package br.com.lucasisrael.jetposemovies.moviesgenre.data.datasource
 
-import br.com.lucasisrael.jetposemovies.moviesgenre.data.respose.MoviesFromGenreResponse
+import br.com.lucasisrael.jetposemovies.moviesgenre.data.response.MoviesFromGenreResponse
 
 interface MoviesFromGenreDataSource {
     suspend fun getMoviesFromGenre(id: String): MoviesFromGenreResponse
