+++ app/src/main/java/br/com/lucasisrael/jetposemovies/moviesgenre/data/datasource/MoviesFromGenreDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package br.com.lucasisrael.jetposemovies.moviesgenre.data.datasource
 
 import br.com.lucasisrael.jetposemovies.moviesgenre.data.api.MoviesGenreApi
-import br.com.lucasisrael.jetposemovies.moviesgenre.data.respose.MoviesFromGenreResponse
+import br.com.lucasisrael.jetposemovies.moviesgenre.data.response.MoviesFromGenreResponse
 import javax.inject.Inject
 
 class MoviesFromGenreDataSourceImpl @Inject constructor(private val api: MoviesGenreApi) :
