+++ app/src/main/java/br/com/lucasisrael/jetposemovies/moviesgenre/data/repository/MoviesFromGenreRepository.kt
@@ -1,7 +1,7 @@
 package br.com.lucasisrael.jetposemovies.moviesgenre.data.repository
 
 import br.com.lucasisrael.jetposemovies.common.models.Resource
-import br.com.lucasisrael.jetposemovies.moviesgenre.data.respose.MoviesFromGenreResponse
+import br.com.lucasisrael.jetposemovies.moviesgenre.data.response.MoviesFromGenreResponse
 
 interface MoviesFromGenreRepository {
     suspend fun getMoviesFromGenre(id: String): Resource<MoviesFromGenreResponse?>
